refactor: 세션 기반 인증을 JWT 기반 인증으로 완전 교체
- 세션 관련 클래스 및 인터페이스 제거 (LoginSession, SessionStore, InMemorySessionStore)
- SessionAuthenticationFilter를 JwtAuthenticationFilter로 교체
- AuthController에서 세션 토큰 검증 로직을 JWT 기반으로 변경
- AuthService의 토큰 검증 및 사용자 조회 로직 단순화
- WebSocket 인증도 JWT 기반으로 통합
- 관련 테스트 코드 업데이트
