+++ src/main/java/com/example/chatapp/infrastructure/auth/SHA256PasswordEncoder.java
